requirements.txt
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
nested-lookup==0.2.21
numpy==1.18.3
pandas==1.0.3
pylint==2.4.4
pylzma==0.5.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.11.2
